bashcompletions, zshcompletions: move to subpackages
