+++ cmd/wksctl/bashcompletions/bashcompletions.go
@@ -1,15 +1,19 @@
-package main
+package bashcompletions
 
 import (
 	"log"
 	"os"
 
 	"github.com/spf13/cobra"
+	"github.com/weaveworks/wksctl/pkg/utilities"
 )
 
-var bashCompletionFilename = "wksctl_bash_completion.sh"
+var (
+	output                 string
+	bashCompletionFilename = "wksctl_bash_completion.sh"
+)
 
-var bashCompletionCmd = &cobra.Command{
+var Cmd = &cobra.Command{
 	Use:   "bash-completions",
 	Short: "Generate bash completion scripts",
 	Long: `To generate completion files, run
@@ -19,20 +23,19 @@ wksctl bash-completions [-o <completions_file|directory>]
 and follow instructions for your OS to configure/install the completion file.
 `,
 	Run: func(cmd *cobra.Command, args []string) {
-		if Output != "" {
-			outfile, err := CreateFile(Output, bashCompletionFilename)
+		if output != "" {
+			outfile, err := utilities.CreateFile(output, bashCompletionFilename)
 			if err != nil {
 				log.Fatal(err)
 			}
-			rootCmd.GenBashCompletion(outfile)
+			cmd.Root().GenBashCompletion(outfile)
 		} else {
-			rootCmd.GenBashCompletion(os.Stdout)
+			cmd.Root().GenBashCompletion(os.Stdout)
 		}
 	}}
 
 func init() {
-	bashCompletionCmd.Flags().StringVarP(
-		&Output, "output", "o", "",
+	Cmd.Flags().StringVarP(
+		&output, "output", "o", "",
 		"completion filename or directory (default stdout)")
-	rootCmd.AddCommand(bashCompletionCmd)
 }
