+++ pkg/utilities/file.go
@@ -1,4 +1,4 @@
-package main
+package utilities
 
 import (
 	"log"
